fix: message parameters are pointer-wide end to end
Clicking a portrait in the who-is-inside strip crashed: the strip posted
the unit's address through SGameMessage::nParam, an int, and the world
client rebuilt the pointer from it - 0xffffffffc9fa2668 in the report
is a heap address after a 32-bit round trip with sign extension.
Widening nParam alone was not enough: every game message also passes
through the UI screen's SUIMessage::nFirst (also int) on its way to the
screen's Lua handler and back, which is where the who-is-inside
realtime update regressed too. Both fields are intptr_t now, the Lua
bridge reads them back full-width through lua_Number (exact for
user-space addresses), and the container posters and the strip send
the full address instead of a truncated token, so
CSelector::UpdateSelection is a plain pointer compare again. Verified
on bldbg2: the portrait click selects the occupant, and the strip
shrinks live as the squad unloads.

+++ Sources/src/Common/MOUnitInfantry.cpp
@@ -381,7 +381,7 @@ void CMOUnitInfantry::SetSquad( interface IMOSquad *_pSquad )
 	if ( pSquad )
 	{
 		pSquad->Load( this, true );
-		GetSingleton<IInput>()->AddMessage( SGameMessage(MC_UPDATE_WHO_IN_CONTAINER, static_cast<int>( reinterpret_cast<std::uintptr_t>( GetContainer() ) )) );
+		GetSingleton<IInput>()->AddMessage( SGameMessage(MC_UPDATE_WHO_IN_CONTAINER, reinterpret_cast<std::intptr_t>( GetContainer() )) );
 	}
 	else
 	{
