+++ Sources/src/Common/UISquadElement.cpp
@@ -324,7 +324,7 @@ bool CUISquadElement::OnLButtonDown( const CVec2 &vPos, EMouseState mouseState )
 			CUIUnitObserver *observer = *(passangers.begin());
 			if ( observer->GetMOUnit()->CanSelect() )
 			{
-				msg.nParam = static_cast<int>( reinterpret_cast<std::uintptr_t>( observer->GetMOUnit() ) );
+				msg.nParam = reinterpret_cast<std::intptr_t>( observer->GetMOUnit() );
 				GetSingleton<IInput>()->AddMessage( msg );
 			}
 		}
@@ -335,7 +335,7 @@ bool CUISquadElement::OnLButtonDown( const CVec2 &vPos, EMouseState mouseState )
 				CUIUnitObserver *observer = *(it);
 				if ( observer->GetMOUnit()->CanSelect() )
 				{
-					msg.nParam = static_cast<int>( reinterpret_cast<std::uintptr_t>( observer->GetMOUnit() ) );
+					msg.nParam = reinterpret_cast<std::intptr_t>( observer->GetMOUnit() );
 					GetSingleton<IInput>()->AddMessage( msg );
 				}
 			}
